app/s3_trigger_file_metadata/requirements.txt
## Requirements for trigger_function_file_metadata goes here.
google-cloud-pubsub